+++ api/build.gradle
@@ -38,6 +38,8 @@ dependencies {
     implementation 'org.modelmapper:modelmapper:2.4.4'
     implementation 'org.apache.poi:poi:5.1.0'
     implementation 'org.apache.poi:poi-ooxml:5.1.0'
+    implementation 'com.univocity:univocity-parsers:2.9.1'
+
 
     // Lombok
     compileOnly 'org.projectlombok:lombok'
