+++ api/src/main/java/com/eleks/academy/pharmagator/parsers/csv/CsvFileParser.java
@@ -1,16 +1,17 @@
-package com.eleks.academy.pharmagator.parser;
+package com.eleks.academy.pharmagator.parsers.csv;
 
+import com.eleks.academy.pharmagator.parsers.ICsvParser;
 import com.univocity.parsers.common.processor.BeanListProcessor;
 import com.univocity.parsers.csv.CsvParser;
 import com.univocity.parsers.csv.CsvParserSettings;
-import org.springframework.stereotype.Component;
+import lombok.AllArgsConstructor;
 
 import java.io.InputStream;
 import java.util.List;
 import java.util.stream.Stream;
 
-@Component
-public class CsvFileParser implements ICsvParser {
+@AllArgsConstructor
+public sealed class CsvFileParser implements ICsvParser permits MedicineRegistryCsvParser{
 
     @Override
     public <T> Stream<T> parse(InputStream inputStream, Class<T> modelType) {
@@ -41,7 +42,9 @@ public class CsvFileParser implements ICsvParser {
 
 
 
-    private CsvParserSettings getCsvParserSettings() {
+
+
+    protected CsvParserSettings getCsvParserSettings() {
         CsvParserSettings parserSettings = new CsvParserSettings();
 
         //the default value is 4096 which is not suitable for parsing
@@ -55,7 +58,7 @@ public class CsvFileParser implements ICsvParser {
         return parserSettings;
     }
 
-    private CsvParser getParser(CsvParserSettings settings) {
+    protected CsvParser getParser(CsvParserSettings settings) {
         return new CsvParser(settings);
     }
